+++ ChangeLog.md
@@ -4,7 +4,7 @@
 All notable changes to this project will be documented in this file.
 
 
-## [1.12.32] - 2024-08-xx
+## [1.12.33] - 2024-08-xx
 
 ### Added
 
@@ -13,6 +13,15 @@ All notable changes to this project will be documented in this file.
 - :openai module support for assistant message api (in work)
 - :openai module support for assistant run api (...)
 - :openai module support for assistant run steps api (...)
+
+
+
+## [1.12.32] - 2024-08-29
+
+### Fixed
+
+- a bug in the :excel module when reading values `(excel/read-val sheet row col)`
+  from string cells with formulas and the cell's value was `nil`.